Change the max size of HTTP POST sent to the GUI
We are trying to upload files to the DQM GUI for testing, however getting error: ERROR HTTP Error 413: Request Entity Too Large

After changing this parameter on the test cluster https://cmsweb-test2.cern.ch/dqm/dev/  we confirmed it now works. Therefore we would like to have this configuration to allow the upload of Root files which can be up to 1 GB but just to be safe 10GB would make the trick.